Fixes surfacing of errors
