+++ packages/vscode-extension/src/contextService.ts
@@ -39,7 +39,7 @@ type SchemaPoller = {
     database?: string,
   ): Promise<ConnnectionResult>;
   disconnect(): void;
-  runQuery(query: string): Promise<QueryResult | 'error'>;
+  runQuery(query: string): Promise<QueryResult | Error>;
 };
 
 let _context: ExtensionContext | undefined;
